Updated Growl.framework to compile with OS X 10.8 as SDK, and added support for Apple's Notification Center.  Logic is as follows:
* If Growl is running, notices go to Growl.
* If Growl is not running and we're on 10.8 or later, notices go to Notification Center.
* If Growl is not running and we're on 10.7 or 10.6, notices go to Mist.

Behavior is as close to normal Growl behavior as could be managed, though can be tuned by global defaults for power users.  By default, non-sticky notices are removed from Notification Center after 120 seconds, while sticky notices remain there.  Clicking on a notice triggers the usual Growl callback and removes the notice from Notification Center.

For power users, there is also a global preference which can be set to make all notices go to Notification Center whether or not Growl is running.  This allows you to use Notification Center for display, but use Growl for automation such as forwarding important notices to an email address.

+++ Core/Source/GrowlGNTPOutgoingPacket.m
@@ -344,11 +344,11 @@
 - (NSString *)description
 {
 	NSMutableString *description = [NSMutableString string];
-	[description appendFormat:@"<%@: %x: \"", NSStringFromClass([self class]), self];
+	[description appendFormat:@"<%@: %x: \"", NSStringFromClass([self class]), (unsigned int)self];
 	
 	for(NSObject<GNTPOutgoingItem> *item in [self outgoingItems]) {
 		if ([item isKindOfClass:[GrowlGNTPBinaryChunk class]]) {
-			[description appendFormat:@"Binary chunk %@, length %d\n",
+			[description appendFormat:@"Binary chunk %@, length %lu\n",
 			 [(GrowlGNTPBinaryChunk *)item identifier], [(GrowlGNTPBinaryChunk *)item length]];
 		} else {
 			NSString *string = [[[NSString alloc] initWithData:[item GNTPRepresentation]
